feat:매출 도메인 추가

+++ src/main/java/com/mcfuturepartners/crm/api/product/dto/ProductRevenueDto.java
@@ -17,13 +17,13 @@ public class ProductRevenueDto{
                                 .getRegDate()
                                 .isAfter(LocalDate.now().atStartOfDay()))
                         .map(order -> order.getProduct().getPrice())
-                        .reduce(0, Integer::sum))
+                        .reduce(0L, Long::sum))
                 .monthlySales(product.getOrders().stream()
                         .filter(order -> order
                                 .getRegDate()
                                 .isAfter(LocalDate.of(LocalDate.now().getYear(),LocalDate.now().getMonth().getValue(),1).atStartOfDay()))
                         .map(order -> order.getProduct().getPrice())
-                        .reduce(0, Integer::sum))
+                        .reduce(0L, Long::sum))
                 .build();
     }
 }
